Changes to code base are as listed below
Separate application configuration from the webserver into a trait that
was included in the webser
Separated route logic into a base class
Introduced basic tests to codebase
