+++ src/main/scala/com/lagos/scalameetup/upnepa/WebServer.scala
@@ -9,14 +9,9 @@ import akka.event.Logging
 import akka.http.scaladsl.Http
 import akka.http.scaladsl.server.Directives._
 import akka.stream.ActorMaterializer
-import com.typesafe.config.ConfigFactory
-import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport._
-import spray.json.DefaultJsonProtocol._
-import spray.json.PrettyPrinter
-
-case class Message(text: String)
-
-object WebServer extends App {
+import com.lagos.scalameetup.upnepa.config.ApplicationConfiguration
+import com.lagos.scalameetup.upnepa.routes.BaseRoutes
+object WebServer extends App with BaseRoutes with ApplicationConfiguration{
   /**
     * Set up needed to provide resources (actor system, actor materializer, and execution context for futures)
     * used internally by akka http
@@ -24,30 +19,6 @@ object WebServer extends App {
   implicit val system = ActorSystem("upnepa-chatbot-system")
   implicit val materializer = ActorMaterializer()
   implicit val executionContext = system.dispatcher
-
-  /**
-    * For marshalling and unmarshalling json to and from scala objects (Message objects)
-    */
-  implicit val printer = PrettyPrinter
-  implicit val messageFormat = jsonFormat1(Message)
-
-  /**
-    * Used to read configuration data from .conf files
-    */
-  val config = ConfigFactory.load()
-  val host =  config.getString("http.host")
-  val port =  config.getInt("http.port")
-
-  /**
-    * http endpoints routes
-    */
-  val route =
-    pathEndOrSingleSlash {
-      get {
-        complete(Message("Hello from up-nepa chatbot server!"))
-      }
-    }
-
   /**
     * Used for logging messages to the console
     */
@@ -59,4 +30,5 @@ object WebServer extends App {
   val bindingFuture = Http().bindAndHandle(route, host, port)
 
   logger.info(s"Server started at http://localhost:$port... \n")
+
 }
