+++ src/main/scala/com/lagos/scalameetup/upnepa/routes/BaseRoutes.scala
@@ -1,0 +1,32 @@
+package com.lagos.scalameetup.upnepa.routes
+
+import akka.http.scaladsl.server.Directives._
+import spray.json.DefaultJsonProtocol._
+import spray.json.PrettyPrinter
+
+/**
+ * Created by olushola on 7/19/17.
+ * Separating routes for easy readibilty
+ */
+trait BaseRoutes {
+
+  case class Message(text: String)
+
+  /**
+   * For marshalling and unmarshalling json to and from scala objects (Message objects)
+   */
+  implicit val printer = PrettyPrinter
+  implicit val messageFormat = jsonFormat1(Message)
+
+  /**
+   * http endpoints routes
+   */
+  val route =
+    pathEndOrSingleSlash {
+      get {
+        complete("Hello from up-nepa chatbot server!")
+      }
+    }
+
+
+}
