+++ src/test/scala/com/lagos/scalameetup/upnepa/WebServerSpec.scala
@@ -1,0 +1,8 @@
+package com.lagos.scalameetup.upnepa
+
+/**
+ * Created by olushola on 7/19/17.
+ */
+class WebServerSpec {
+
+}
